Split out account type-related actions
