+++ client/state/login/actions/index.js
@@ -1,20 +1,7 @@
 /**
- * External dependencies
- */
-import { defer } from 'lodash';
-import { translate } from 'i18n-calypso';
-
-/**
  * Internal dependencies
  */
-import {
-	LOGIN_AUTH_ACCOUNT_TYPE_REQUEST,
-	LOGIN_AUTH_ACCOUNT_TYPE_REQUESTING,
-	LOGIN_AUTH_ACCOUNT_TYPE_REQUEST_FAILURE,
-	LOGIN_AUTH_ACCOUNT_TYPE_RESET,
-	LOGIN_FORM_UPDATE,
-} from 'state/action-types';
-import { recordTracksEventWithClientId as recordTracksEvent } from 'state/analytics/actions';
+import { LOGIN_FORM_UPDATE } from 'state/action-types';
 import 'state/data-layer/wpcom/login-2fa';
 import 'state/data-layer/wpcom/users/auth-options';
 
@@ -35,58 +22,6 @@ export {
 	startPollAppPushAuth,
 	stopPollAppPushAuth,
 } from 'state/login/actions/push';
+export { getAuthAccountType, resetAuthAccountType } from 'state/login/actions/auth-account-type.js';
 
 export const formUpdate = () => ( { type: LOGIN_FORM_UPDATE } );
-
-/**
- * Retrieves the type of authentication of the account (regular, passwordless ...) of the specified user.
- *
- * @param  {string}   usernameOrEmail Identifier of the user
- * @returns {Function}                 A thunk that can be dispatched
- */
-export const getAuthAccountType = usernameOrEmail => dispatch => {
-	dispatch( recordTracksEvent( 'calypso_login_block_login_form_get_auth_type' ) );
-
-	dispatch( {
-		type: LOGIN_AUTH_ACCOUNT_TYPE_REQUEST,
-		usernameOrEmail,
-	} );
-
-	if ( usernameOrEmail === '' ) {
-		const error = {
-			code: 'empty_username',
-			message: translate( 'Please enter a username or email address.' ),
-			field: 'usernameOrEmail',
-		};
-
-		dispatch(
-			recordTracksEvent( 'calypso_login_block_login_form_get_auth_type_failure', {
-				error_code: error.code,
-				error_message: error.message,
-			} )
-		);
-
-		defer( () => {
-			dispatch( {
-				type: LOGIN_AUTH_ACCOUNT_TYPE_REQUEST_FAILURE,
-				error,
-			} );
-		} );
-
-		return;
-	}
-
-	dispatch( {
-		type: LOGIN_AUTH_ACCOUNT_TYPE_REQUESTING,
-		usernameOrEmail,
-	} );
-};
-
-/**
- * Resets the type of authentication of the account of the current user.
- *
- * @returns {object} An action that can be dispatched
- */
-export const resetAuthAccountType = () => ( {
-	type: LOGIN_AUTH_ACCOUNT_TYPE_RESET,
-} );
